chore: separating app-routing components
