+++ lib/app/view/app.dart
@@ -1,14 +1,17 @@
 import 'package:ax_dapp/app/bloc/app_bloc.dart';
-import 'package:ax_dapp/debug/views/debug_app_wrapper.dart';
-import 'package:ax_dapp/landing_page/landing_page.dart';
+import 'package:ax_dapp/app/view/app_routing.dart';
+import 'package:ax_dapp/league/league_search/bloc/league_bloc.dart';
+import 'package:ax_dapp/league/repository/prize_pool_repository.dart';
+import 'package:ax_dapp/league/usecases/league_use_case.dart';
 import 'package:ax_dapp/service/tracking/tracking_cubit.dart';
 import 'package:ax_dapp/wallet/wallet.dart';
 import 'package:config_repository/config_repository.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:league_repository/league_repository.dart';
 import 'package:tokens_repository/tokens_repository.dart';
 import 'package:tracking_repository/tracking_repository.dart';
+import 'package:use_cases/stream_app_data_changes_use_case.dart';
 import 'package:wallet_repository/wallet_repository.dart';
 
 class App extends StatelessWidget {
@@ -41,37 +44,17 @@ class App extends StatelessWidget {
           create: (context) => TrackingCubit(
             context.read<TrackingRepository>(),
           )..setup(),
-        )
+        ),
+        BlocProvider(
+          create: (context) => LeagueBloc(
+            leagueRepository: context.read<LeagueRepository>(),
+            streamAppDataChanges: context.read<StreamAppDataChangesUseCase>(),
+            prizePoolRepository: context.read<PrizePoolRepository>(),
+            leagueUseCase: context.read<LeagueUseCase>(),
+          ),
+        ),
       ],
-      child: kDebugMode
-          ? const DebugAppWrapper(
-              home: _MaterialApp(),
-            )
-          : const _MaterialApp(),
-    );
-  }
-}
-
-class _MaterialApp extends StatelessWidget {
-  // ignore: use_super_parameters
-  const _MaterialApp({
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'AthleteX',
-      debugShowCheckedModeBanner: false,
-      initialRoute: '/',
-      theme: ThemeData(
-        canvasColor: Colors.transparent,
-        brightness: Brightness.dark,
-        primaryColor: Colors.yellow[700],
-        colorScheme: ColorScheme.fromSwatch(brightness: Brightness.dark)
-            .copyWith(secondary: Colors.black),
-      ),
-      home: const LandingPage(),
+      child: const EntryApp(),
     );
   }
 }
