+++ lib/app/widgets/bottom_navigation_bar_mobile.dart
@@ -1,25 +1,46 @@
 import 'package:flutter/material.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+// TODO(Ryan): re-enable this import for the league feature
+// import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:go_router/go_router.dart';
 
-class BottomNavigationBarMobile extends StatelessWidget {
+class BottomNavigationBarMobile extends StatefulWidget {
   const BottomNavigationBarMobile({
     super.key,
-    required this.selectedIndex,
   });
 
-  final int selectedIndex;
+  @override
+  State<BottomNavigationBarMobile> createState() =>
+      _BottomNavigationBarMobileState();
+}
+
+class _BottomNavigationBarMobileState extends State<BottomNavigationBarMobile> {
+  int _selectedIndex = 0;
+  void _onItemTapped(int index) {
+    setState(() {
+      _selectedIndex = index;
+    });
+
+    switch (index) {
+      case 0:
+        context.goNamed('scout');
+        break;
+      case 1:
+        context.goNamed('trade');
+        break;
+      case 2:
+        context.goNamed('pool');
+        break;
+      case 3:
+        context.goNamed('farm');
+        break;
+      case 4:
+        context.goNamed('league');
+        break;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
-    Color iconColor(int index) {
-      if (index == selectedIndex) {
-        return Colors.white;
-      } else {
-        return Colors.grey;
-      }
-    }
-
     return BottomNavigationBar(
       showUnselectedLabels: true,
       selectedLabelStyle: const TextStyle(
@@ -38,7 +59,6 @@ class BottomNavigationBarMobile extends StatelessWidget {
             'assets/images/search.png',
             height: 24,
             width: 24,
-            color: iconColor(0),
           ),
           label: 'Scout',
         ),
@@ -47,7 +67,6 @@ class BottomNavigationBarMobile extends StatelessWidget {
             'assets/images/swap.png',
             height: 24,
             width: 24,
-            color: iconColor(1),
           ),
           label: 'Trade',
         ),
@@ -56,7 +75,6 @@ class BottomNavigationBarMobile extends StatelessWidget {
             'assets/images/coins.png',
             height: 24,
             width: 24,
-            color: iconColor(2),
           ),
           label: 'Pool',
         ),
@@ -65,7 +83,6 @@ class BottomNavigationBarMobile extends StatelessWidget {
             'assets/images/barn.png',
             height: 24,
             width: 24,
-            color: iconColor(3),
           ),
           label: 'Farm',
         ),
@@ -74,33 +91,14 @@ class BottomNavigationBarMobile extends StatelessWidget {
         //   icon: FaIcon(
         //     FontAwesomeIcons.trophy,
         //     size: 24,
-        //     color: iconColor(4),
         //   ),
         //   label: 'League',
         // ),
       ],
-      currentIndex: selectedIndex,
+      currentIndex: _selectedIndex,
       selectedItemColor: Colors.white,
       unselectedItemColor: Colors.grey,
-      onTap: (index) {
-        switch (index) {
-          case 0:
-            context.goNamed('scout');
-            break;
-          case 1:
-            context.goNamed('trade');
-            break;
-          case 2:
-            context.goNamed('pool');
-            break;
-          case 3:
-            context.goNamed('farm');
-            break;
-          case 4:
-            context.goNamed('league');
-            break;
-        }
-      },
+      onTap: _onItemTapped,
     );
   }
 }
