+++ lib/league/league_draft/views/desktop_league_draft.dart
@@ -40,74 +40,242 @@ class DesktopLeagueDraft extends StatelessWidget {
     const bookValueHeadingWR = 0.1;
     const myTeamHeadingWR = 0.35;
     const athleteCountHeadingWR = 0.153;
-    return global.buildPage(
-      context,
-      BlocBuilder<LeagueDraftBloc, LeagueDraftState>(
-        builder: (context, state) {
-          final bloc = context.read<LeagueDraftBloc>();
-          return LayoutBuilder(
-            builder: (BuildContext context, BoxConstraints constraints) {
-              final width = constraints.maxWidth;
-              final height = constraints.maxHeight;
-              if (width <= 768) {
-                textSize = 12.0;
-                leagueHeaderTextSize = 18.0;
-                verticalView = false;
-                athleteColumnTitle = 'Athlete';
-                bookValueTitle = 'Book Value';
-              }
-              return Align(
-                child: Container(
-                  height: height * 0.75,
-                  width: width * 0.90,
-                  clipBehavior: Clip.hardEdge,
-                  decoration: BoxDecoration(
-                    border: Border.all(
-                      color: Colors.transparent,
-                    ),
+    return BlocBuilder<LeagueDraftBloc, LeagueDraftState>(
+      builder: (context, state) {
+        final bloc = context.read<LeagueDraftBloc>();
+        return LayoutBuilder(
+          builder: (BuildContext context, BoxConstraints constraints) {
+            final width = constraints.maxWidth;
+            final height = constraints.maxHeight;
+            if (width <= 768) {
+              textSize = 12.0;
+              leagueHeaderTextSize = 18.0;
+              verticalView = false;
+              athleteColumnTitle = 'Athlete';
+              bookValueTitle = 'Book Value';
+            }
+            return Align(
+              child: Container(
+                height: height * 0.75,
+                width: width * 0.90,
+                clipBehavior: Clip.hardEdge,
+                decoration: BoxDecoration(
+                  border: Border.all(
+                    color: Colors.transparent,
                   ),
-                  child: Column(
-                    children: <Widget>[
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          SizedBox(
-                            width: 70,
-                            child: TextButton(
-                              onPressed: () {
-                                Navigator.pop(context);
-                              },
-                              child: const Icon(
-                                Icons.arrow_back,
-                                size: 50,
-                                color: Colors.white,
+                ),
+                child: Column(
+                  children: <Widget>[
+                    Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        SizedBox(
+                          width: 70,
+                          child: TextButton(
+                            onPressed: () {
+                              Navigator.pop(context);
+                            },
+                            child: const Icon(
+                              Icons.arrow_back,
+                              size: 50,
+                              color: Colors.white,
+                            ),
+                          ),
+                        ),
+                        Text(
+                          league.name,
+                          style: TextStyle(
+                            color: const Color(0xFFFEC500),
+                            fontFamily: 'OpenSans',
+                            fontSize: leagueHeaderTextSize,
+                            fontWeight: FontWeight.w400,
+                          ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.all(8),
+                          child: TextButton(
+                            style: ButtonStyle(
+                              shape: MaterialStateProperty.all<
+                                  RoundedRectangleBorder>(
+                                RoundedRectangleBorder(
+                                  borderRadius: BorderRadius.circular(100),
+                                  side: BorderSide(color: Colors.amber[400]!),
+                                ),
+                              ),
+                            ),
+                            onPressed: () => {context.goNamed('scout')},
+                            child: Text(
+                              'Go Buy APTs',
+                              style: TextStyle(
+                                color: const Color(0xFFFEC500),
+                                fontFamily: 'OpenSans',
+                                fontSize: textSize,
+                                fontWeight: FontWeight.w400,
                               ),
                             ),
                           ),
-                          Text(
-                            league.name,
-                            style: TextStyle(
-                              color: const Color(0xFFFEC500),
-                              fontFamily: 'OpenSans',
-                              fontSize: leagueHeaderTextSize,
-                              fontWeight: FontWeight.w400,
+                        ),
+                      ],
+                    ),
+                    const SizedBox(
+                      height: 50,
+                    ),
+                    Padding(
+                      padding: verticalView
+                          ? const EdgeInsets.symmetric(horizontal: 30)
+                          : EdgeInsets.zero,
+                      child: Row(
+                        children: [
+                          FittedBox(
+                            child: SizedBox(
+                              width: width * athleteHeadingWR,
+                              child: Text(
+                                athleteColumnTitle,
+                                textAlign: TextAlign.center,
+                                style: TextStyle(fontSize: textSize),
+                              ),
                             ),
                           ),
-                          Padding(
-                            padding: const EdgeInsets.all(8),
-                            child: TextButton(
-                              style: ButtonStyle(
-                                shape: MaterialStateProperty.all<
-                                    RoundedRectangleBorder>(
-                                  RoundedRectangleBorder(
-                                    borderRadius: BorderRadius.circular(100),
-                                    side: BorderSide(color: Colors.amber[400]!),
+                          FittedBox(
+                            child: SizedBox(
+                              width: width * bookValueHeadingWR,
+                              child: Text(
+                                bookValueTitle,
+                                textAlign: TextAlign.center,
+                                style: TextStyle(fontSize: textSize),
+                              ),
+                            ),
+                          ),
+                          FittedBox(
+                            child: SizedBox(
+                              width: width * myTeamHeadingWR,
+                              child: Text(
+                                'My Team',
+                                textAlign: TextAlign.end,
+                                style: TextStyle(fontSize: textSize),
+                              ),
+                            ),
+                          ),
+                          FittedBox(
+                            child: SizedBox(
+                              width: width * athleteCountHeadingWR,
+                              child: Text(
+                                '${state.athleteCount} / ${league.teamSize}',
+                                textAlign: TextAlign.end,
+                                style: TextStyle(fontSize: textSize),
+                              ),
+                            ),
+                          ),
+                        ],
+                      ),
+                    ),
+                    Expanded(
+                      child: DecoratedBox(
+                        decoration: const BoxDecoration(
+                          border: Border(
+                            top: BorderSide(
+                              color: Color(0xFF646464),
+                              width: 2,
+                            ),
+                          ),
+                        ),
+                        child: Flex(
+                          direction:
+                              verticalView ? Axis.horizontal : Axis.vertical,
+                          children: [
+                            Expanded(
+                              child: Material(
+                                child: state.status == BlocStatus.loading
+                                    ? const Loader()
+                                    : Align(
+                                        alignment: Alignment.topCenter,
+                                        child: ListView.builder(
+                                          padding: EdgeInsets.zero,
+                                          shrinkWrap: true,
+                                          itemCount: ownedApts.length,
+                                          itemBuilder: (context, index) {
+                                            return APTCard(
+                                              apt: ownedApts[index],
+                                              teamSize: league.teamSize,
+                                              height: height,
+                                            );
+                                          },
+                                        ),
+                                      ),
+                              ),
+                            ),
+                            if (verticalView)
+                              const VerticalDivider()
+                            else
+                              const Divider(
+                                thickness: 3,
+                              ),
+                            Expanded(
+                              child: Material(
+                                child: Align(
+                                  alignment: Alignment.topCenter,
+                                  child: ListView.builder(
+                                    padding: EdgeInsets.zero,
+                                    shrinkWrap: true,
+                                    itemCount: myAptTeam.length,
+                                    itemBuilder: (context, index) {
+                                      return MyTeamCard(
+                                        apt: myAptTeam[index],
+                                        height: height,
+                                      );
+                                    },
                                   ),
                                 ),
                               ),
-                              onPressed: () => {context.goNamed('scout')},
+                            ),
+                          ],
+                        ),
+                      ),
+                    ),
+                    Align(
+                      alignment: verticalView
+                          ? Alignment.centerRight
+                          : Alignment.center,
+                      child: Padding(
+                        padding: verticalView
+                            ? EdgeInsets.only(right: width / 5)
+                            : EdgeInsets.zero,
+                        child: DecoratedBox(
+                          decoration: BoxDecoration(
+                            color: const Color(0xFF3C3009),
+                            borderRadius: BorderRadius.circular(100),
+                            border: Border.all(
+                              color: Colors.transparent,
+                            ),
+                          ),
+                          child: Padding(
+                            padding: const EdgeInsets.all(8),
+                            child: TextButton(
+                              onPressed: () {
+                                if (isWalletConnected && myAptTeam.isNotEmpty) {
+                                  bloc.add(
+                                    ConfirmTeam(
+                                      walletAddress: walletAddress,
+                                      leagueID: league.leagueID,
+                                      myTeam: myAptTeam,
+                                      existingTeam: existingTeam,
+                                      prizePoolAddress: league.prizePoolAddress,
+                                      entryFee: league.entryFee,
+                                    ),
+                                  );
+                                  Navigator.pop(context);
+                                } else if (myAptTeam.isEmpty) {
+                                  context.showWarningToast(
+                                    title: 'No Athletes Selected!',
+                                    description:
+                                        'Please Add Athletes Before Entering A League!',
+                                  );
+                                } else {
+                                  context.showWalletWarningToast();
+                                }
+                              },
                               child: Text(
-                                'Go Buy APTs',
+                                'Confirm',
                                 style: TextStyle(
                                   color: const Color(0xFFFEC500),
                                   fontFamily: 'OpenSans',
@@ -117,189 +285,16 @@ class DesktopLeagueDraft extends StatelessWidget {
                               ),
                             ),
                           ),
-                        ],
-                      ),
-                      const SizedBox(
-                        height: 50,
-                      ),
-                      Padding(
-                        padding: verticalView
-                            ? const EdgeInsets.symmetric(horizontal: 30)
-                            : EdgeInsets.zero,
-                        child: Row(
-                          children: [
-                            FittedBox(
-                              child: SizedBox(
-                                width: width * athleteHeadingWR,
-                                child: Text(
-                                  athleteColumnTitle,
-                                  textAlign: TextAlign.center,
-                                  style: TextStyle(fontSize: textSize),
-                                ),
-                              ),
-                            ),
-                            FittedBox(
-                              child: SizedBox(
-                                width: width * bookValueHeadingWR,
-                                child: Text(
-                                  bookValueTitle,
-                                  textAlign: TextAlign.center,
-                                  style: TextStyle(fontSize: textSize),
-                                ),
-                              ),
-                            ),
-                            FittedBox(
-                              child: SizedBox(
-                                width: width * myTeamHeadingWR,
-                                child: Text(
-                                  'My Team',
-                                  textAlign: TextAlign.end,
-                                  style: TextStyle(fontSize: textSize),
-                                ),
-                              ),
-                            ),
-                            FittedBox(
-                              child: SizedBox(
-                                width: width * athleteCountHeadingWR,
-                                child: Text(
-                                  '${state.athleteCount} / ${league.teamSize}',
-                                  textAlign: TextAlign.end,
-                                  style: TextStyle(fontSize: textSize),
-                                ),
-                              ),
-                            ),
-                          ],
                         ),
                       ),
-                      Expanded(
-                        child: DecoratedBox(
-                          decoration: const BoxDecoration(
-                            border: Border(
-                              top: BorderSide(
-                                color: Color(0xFF646464),
-                                width: 2,
-                              ),
-                            ),
-                          ),
-                          child: Flex(
-                            direction:
-                                verticalView ? Axis.horizontal : Axis.vertical,
-                            children: [
-                              Expanded(
-                                child: Material(
-                                  child: state.status == BlocStatus.loading
-                                      ? const Loader()
-                                      : Align(
-                                          alignment: Alignment.topCenter,
-                                          child: ListView.builder(
-                                            padding: EdgeInsets.zero,
-                                            shrinkWrap: true,
-                                            itemCount: ownedApts.length,
-                                            itemBuilder: (context, index) {
-                                              return APTCard(
-                                                apt: ownedApts[index],
-                                                teamSize: league.teamSize,
-                                                height: height,
-                                              );
-                                            },
-                                          ),
-                                        ),
-                                ),
-                              ),
-                              if (verticalView)
-                                const VerticalDivider()
-                              else
-                                const Divider(
-                                  thickness: 3,
-                                ),
-                              Expanded(
-                                child: Material(
-                                  child: Align(
-                                    alignment: Alignment.topCenter,
-                                    child: ListView.builder(
-                                      padding: EdgeInsets.zero,
-                                      shrinkWrap: true,
-                                      itemCount: myAptTeam.length,
-                                      itemBuilder: (context, index) {
-                                        return MyTeamCard(
-                                          apt: myAptTeam[index],
-                                          height: height,
-                                        );
-                                      },
-                                    ),
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                      Align(
-                        alignment: verticalView
-                            ? Alignment.centerRight
-                            : Alignment.center,
-                        child: Padding(
-                          padding: verticalView
-                              ? EdgeInsets.only(right: width / 5)
-                              : EdgeInsets.zero,
-                          child: DecoratedBox(
-                            decoration: BoxDecoration(
-                              color: const Color(0xFF3C3009),
-                              borderRadius: BorderRadius.circular(100),
-                              border: Border.all(
-                                color: Colors.transparent,
-                              ),
-                            ),
-                            child: Padding(
-                              padding: const EdgeInsets.all(8),
-                              child: TextButton(
-                                onPressed: () {
-                                  if (isWalletConnected &&
-                                      myAptTeam.isNotEmpty) {
-                                    bloc.add(
-                                      ConfirmTeam(
-                                        walletAddress: walletAddress,
-                                        leagueID: league.leagueID,
-                                        myTeam: myAptTeam,
-                                        existingTeam: existingTeam,
-                                        prizePoolAddress:
-                                            league.prizePoolAddress,
-                                        entryFee: league.entryFee,
-                                      ),
-                                    );
-                                    Navigator.pop(context);
-                                  } else if (myAptTeam.isEmpty) {
-                                    context.showWarningToast(
-                                      title: 'No Athletes Selected!',
-                                      description:
-                                          'Please Add Athletes Before Entering A League!',
-                                    );
-                                  } else {
-                                    context.showWalletWarningToast();
-                                  }
-                                },
-                                child: Text(
-                                  'Confirm',
-                                  style: TextStyle(
-                                    color: const Color(0xFFFEC500),
-                                    fontFamily: 'OpenSans',
-                                    fontSize: textSize,
-                                    fontWeight: FontWeight.w400,
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ),
-                        ),
-                      ),
-                    ],
-                  ),
+                    ),
+                  ],
                 ),
-              );
-            },
-          );
-        },
-      ),
+              ),
+            );
+          },
+        );
+      },
     );
   }
 }
