+++ lib/pool/view/desktop_pool.dart
@@ -52,150 +52,149 @@ class _DesktopPoolState extends State<DesktopPool> {
     final layoutWidth = isWeb ? _width * 0.8 : _width * 0.9;
     final toggleWidth = isWeb ? 260.0 : layoutWidth;
 
-    return global.buildPage(
-      context,
-      SingleChildScrollView(
-        physics: const ClampingScrollPhysics(),
-        child: Container(
-          width: layoutWidth,
-          height: _height - AppBar().preferredSize.height - 10,
-          // Top margin of Pool section is equal to height + 1 of AppBar on
-          // mobile only
-          margin: EdgeInsets.only(top: AppBar().preferredSize.height + 10),
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Container(
-                alignment: Alignment.center,
-                child: Container(
-                  width: toggleWidth,
-                  height: isWeb ? 40 : layoutHeight * 0.06,
-                  margin: EdgeInsets.symmetric(vertical: layoutHeight * 0.01),
-                  decoration: boxDecoration(
-                    Colors.grey[900]!,
-                    100,
-                    1,
-                    Colors.grey[400]!,
-                  ),
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceAround,
-                    children: [
-                      Container(
-                        width: isWeb ? 120 : (toggleWidth / 2) - 5,
-                        decoration: currentTabIndex == 0
-                            ? boxDecoration(
-                                Colors.grey[600]!,
-                                100,
-                                0,
-                                Colors.transparent,
-                              )
-                            : boxDecoration(
-                                Colors.transparent,
-                                100,
-                                0,
-                                Colors.transparent,
-                              ),
-                        child: TextButton(
-                          onPressed: () {
-                            setState(() {
-                              currentTabIndex = 0;
-                            });
-                          },
-                          child: FittedBox(
-                            child: SizedBox(
-                              child: Text(
-                                'Add Liquidity',
-                                style: textStyle(
-                                  Colors.white,
-                                  16,
-                                  isBold: true,
-                                  isUline: false,
-                                ),
+    return SingleChildScrollView(
+      physics: const ClampingScrollPhysics(),
+      child: Container(
+        width: layoutWidth,
+        height: _height - AppBar().preferredSize.height - 10,
+        // Top margin of Pool section is equal to height + 1 of AppBar on
+        // mobile only
+        margin: EdgeInsets.only(top: AppBar().preferredSize.height + 10),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Container(
+              alignment: Alignment.center,
+              child: Container(
+                width: toggleWidth,
+                height: isWeb ? 40 : layoutHeight * 0.06,
+                margin: EdgeInsets.symmetric(vertical: layoutHeight * 0.01),
+                decoration: boxDecoration(
+                  Colors.grey[900]!,
+                  100,
+                  1,
+                  Colors.grey[400]!,
+                ),
+                child: Row(
+                  mainAxisAlignment: MainAxisAlignment.spaceAround,
+                  children: [
+                    Container(
+                      width: isWeb ? 120 : (toggleWidth / 2) - 5,
+                      decoration: currentTabIndex == 0
+                          ? boxDecoration(
+                              Colors.grey[600]!,
+                              100,
+                              0,
+                              Colors.transparent,
+                            )
+                          : boxDecoration(
+                              Colors.transparent,
+                              100,
+                              0,
+                              Colors.transparent,
+                            ),
+                      child: TextButton(
+                        onPressed: () {
+                          setState(() {
+                            currentTabIndex = 0;
+                          });
+                        },
+                        child: FittedBox(
+                          child: SizedBox(
+                            child: Text(
+                              'Add Liquidity',
+                              style: textStyle(
+                                Colors.white,
+                                16,
+                                isBold: true,
+                                isUline: false,
                               ),
                             ),
                           ),
                         ),
                       ),
-                      Container(
-                        width: isWeb ? 120 : (toggleWidth / 2) - 5,
-                        decoration: (currentTabIndex == 1)
-                            ? boxDecoration(
-                                Colors.grey[600]!,
-                                100,
-                                0,
-                                Colors.transparent,
-                              )
-                            : boxDecoration(
-                                Colors.transparent,
-                                100,
-                                0,
-                                Colors.transparent,
-                              ),
-                        child: TextButton(
-                          onPressed: () {
-                            setState(() {
-                              currentTabIndex = 1;
-                            });
-                          },
-                          child: FittedBox(
-                            child: SizedBox(
-                              child: Text(
-                                'My Liquidity',
-                                style: textStyle(
-                                  Colors.white,
-                                  16,
-                                  isBold: true,
-                                  isUline: false,
-                                ),
+                    ),
+                    Container(
+                      width: isWeb ? 120 : (toggleWidth / 2) - 5,
+                      decoration: (currentTabIndex == 1)
+                          ? boxDecoration(
+                              Colors.grey[600]!,
+                              100,
+                              0,
+                              Colors.transparent,
+                            )
+                          : boxDecoration(
+                              Colors.transparent,
+                              100,
+                              0,
+                              Colors.transparent,
+                            ),
+                      child: TextButton(
+                        onPressed: () {
+                          setState(() {
+                            currentTabIndex = 1;
+                          });
+                        },
+                        child: FittedBox(
+                          child: SizedBox(
+                            child: Text(
+                              'My Liquidity',
+                              style: textStyle(
+                                Colors.white,
+                                16,
+                                isBold: true,
+                                isUline: false,
                               ),
                             ),
                           ),
                         ),
                       ),
-                    ],
-                  ),
+                    ),
+                  ],
                 ),
               ),
-              IndexedStack(
-                alignment: AlignmentDirectional.center,
-                index: currentTabIndex,
-                children: [
-                  SizedBox(
-                    height: layoutHeight,
-                    child: BlocProvider(
-                      create: (BuildContext context) => AddLiquidityBloc(
-                        walletRepository: context.read<WalletRepository>(),
-                        tokensRepository: context.read<TokensRepository>(),
-                        streamAppDataChanges:
-                            context.read<StreamAppDataChangesUseCase>(),
-                        repo:
-                            RepositoryProvider.of<GetPoolInfoUseCase>(context),
-                        getAllLiquidityInfoUseCase: RepositoryProvider.of<GetAllLiquidityInfoUseCase>(context),
-                        poolRepository: context.read<PoolRepository>(),
-                      ),
-                      child: AddLiquidityPage(
-                        token0: token0,
-                        token1: token1,
-                      ),
-                    ),
-                  ),
-                  BlocProvider(
-                    create: (_) => MyLiquidityBloc(
+            ),
+            IndexedStack(
+              alignment: AlignmentDirectional.center,
+              index: currentTabIndex,
+              children: [
+                SizedBox(
+                  height: layoutHeight,
+                  child: BlocProvider(
+                    create: (BuildContext context) => AddLiquidityBloc(
                       walletRepository: context.read<WalletRepository>(),
+                      tokensRepository: context.read<TokensRepository>(),
                       streamAppDataChanges:
                           context.read<StreamAppDataChangesUseCase>(),
-                      repo: RepositoryProvider.of<GetAllLiquidityInfoUseCase>(
+                      repo: RepositoryProvider.of<GetPoolInfoUseCase>(context),
+                      getAllLiquidityInfoUseCase:
+                          RepositoryProvider.of<GetAllLiquidityInfoUseCase>(
                         context,
                       ),
+                      poolRepository: context.read<PoolRepository>(),
                     ),
-                    child: MyLiquidityPage(
-                      togglePool: togglePool,
+                    child: AddLiquidityPage(
+                      token0: token0,
+                      token1: token1,
                     ),
                   ),
-                ],
-              )
-            ],
-          ),
+                ),
+                BlocProvider(
+                  create: (_) => MyLiquidityBloc(
+                    walletRepository: context.read<WalletRepository>(),
+                    streamAppDataChanges:
+                        context.read<StreamAppDataChangesUseCase>(),
+                    repo: RepositoryProvider.of<GetAllLiquidityInfoUseCase>(
+                      context,
+                    ),
+                  ),
+                  child: MyLiquidityPage(
+                    togglePool: togglePool,
+                  ),
+                ),
+              ],
+            )
+          ],
         ),
       ),
     );
