+++ lib/wallet/widgets/dialogs/wallet_dialog.dart
@@ -1,4 +1,3 @@
-import 'package:ax_dapp/connect_wallet/mobile_login_page.dart';
 import 'package:ax_dapp/dialogs/promo/connected_wallet_promo_dialog.dart';
 import 'package:ax_dapp/wallet/bloc/wallet_bloc.dart';
 import 'package:ax_dapp/wallet/widgets/dialogs/connect_metamask_dialog.dart';
@@ -133,39 +132,39 @@ class WalletDialog extends StatelessWidget {
                     ),
                   ],
                 ),
-                Visibility(
-                  visible: !kIsWeb,
-                  child: Container(
-                    margin: const EdgeInsets.symmetric(vertical: 4),
-                    width: constraints.maxWidth < 450
-                        ? constraints.maxWidth * 0.62
-                        : constraints.maxWidth * 0.22,
-                    height: 45,
-                    decoration: BoxDecoration(
-                      color: Colors.transparent,
-                      borderRadius: BorderRadius.circular(100),
-                      border: Border.all(color: Colors.grey[400]!, width: 2),
-                    ),
-                    child: TextButton(
-                      onPressed: () {
-                        Navigator.push(
-                          context,
-                          MaterialPageRoute<void>(
-                            builder: (context) => const MobileLoginPage(),
-                          ),
-                        );
-                      },
-                      child: const Text(
-                        'Add/Create wallet',
-                        textAlign: TextAlign.center,
-                        style: TextStyle(
-                          color: Colors.white,
-                          fontFamily: 'OpenSans',
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
+                // Visibility(
+                //   visible: !kIsWeb,
+                //   child: Container(
+                //     margin: const EdgeInsets.symmetric(vertical: 4),
+                //     width: constraints.maxWidth < 450
+                //         ? constraints.maxWidth * 0.62
+                //         : constraints.maxWidth * 0.22,
+                //     height: 45,
+                //     decoration: BoxDecoration(
+                //       color: Colors.transparent,
+                //       borderRadius: BorderRadius.circular(100),
+                //       border: Border.all(color: Colors.grey[400]!, width: 2),
+                //     ),
+                //     child: TextButton(
+                //       onPressed: () {
+                //         Navigator.push(
+                //           context,
+                //           MaterialPageRoute<void>(
+                //             builder: (context) => const MobileLoginPage(),
+                //           ),
+                //         );
+                //       },
+                //       child: const Text(
+                //         'Add/Create wallet',
+                //         textAlign: TextAlign.center,
+                //         style: TextStyle(
+                //           color: Colors.white,
+                //           fontFamily: 'OpenSans',
+                //         ),
+                //       ),
+                //     ),
+                //   ),
+                // ),
               ],
             ),
           ),
